Conflicts:
	main/samba/ChangeLog

+++ main/samba/src/EBox/Samba/Provision.pm
@@ -30,11 +30,14 @@ use EBox::Global;
 use EBox::Users::User;
 use EBox::Users::Group;
 
+use EBox::Users::NamingContext;
+use EBox::Samba::NamingContext;
+
 use Net::DNS;
 use Net::NTP qw(get_ntp_response);
 use Net::Ping;
 use Net::LDAP;
-use Net::LDAP::Util qw(ldap_explode_dn);
+use Net::LDAP::Util qw(ldap_explode_dn canonical_dn);
 use File::Temp qw( tempfile tempdir );
 use File::Slurp;
 use Time::HiRes;
@@ -229,7 +232,7 @@ sub _domainsIP
     my @ipIds = @{$domainIPsModel->ids()};
 
     my $network = EBox::Global->modInstance('network');
-    my $ifaces = $samba->sambaInterfaces();
+    my $ifaces = $network->allIfaces();
 
     my %domainsIp;
     foreach my $iface (@{$ifaces}) {
@@ -280,6 +283,7 @@ sub setupDNS
 {
     my ($self) = @_;
 
+    EBox::info("Setting up DNS");
     my $samba = EBox::Global->modInstance('samba');
 
     if (EBox::Sudo::fileTest('-f', $samba->SAMBA_DNS_KEYTAB())) {
@@ -360,9 +364,82 @@ sub resetSysvolACL
     EBox::Sudo::rootWithoutException($cmd);
 }
 
+sub linkContainer
+{
+    my ($self, $ldb, $ldap) = @_;
+
+    my $usersMod = EBox::Global->modInstance('users');
+    my $sambaMod = EBox::Global->modInstance('samba');
+    my %ldb = %{$ldb};
+    my %ldap = %{$ldap};
+
+    EBox::info("Linking '$ldb{dn}' and '$ldap{dn}'");
+    my $ldapObject = $usersMod->objectFromDN($ldap{dn});
+    my $ldbObject = $sambaMod->objectFromDN($ldb{dn});
+    unless (defined $ldbObject and $ldbObject->exists()) {
+        if ($ldb{create}) {
+            my $dn = ldap_explode_dn($ldb{dn});
+            my $rdn = shift(@{$dn});
+            unless (exists $rdn->{OU}) {
+                throw EBox::Exceptions::Internal("Unable to parse DN $ldb{dn}");
+            }
+            my $name = $rdn->{OU};
+            my $parent = new EBox::Samba::NamingContext(dn => canonical_dn($dn));
+            $ldbObject = EBox::Samba::OU->create(name => $name, parent => $parent);
+            unless (defined $ldbObject and $ldbObject->exists()) {
+                throw EBox::Exceptions::Internal("Unable to create $ldb{dn}");
+            }
+        } else {
+            throw EBox::Exceptions::Internal("Unable to find $ldb{dn} on LDB.");
+        }
+    }
+    unless (defined $ldapObject and $ldapObject->exists()) {
+        if ($ldap{create}) {
+            my $dn = ldap_explode_dn($ldap{dn});
+            my $rdn = shift(@{$dn});
+            unless (exists $rdn->{OU}) {
+                throw EBox::Exceptions::Internal("Unable to parse DN $ldap{dn}");
+            }
+            my $name = $rdn->{OU};
+            my $parent = new EBox::Users::NamingContext(dn => canonical_dn($dn));
+            $ldapObject = EBox::Users::OU->create(name => $name, parent => $parent, ignoreMods=>['samba']);
+            unless (defined $ldapObject and $ldapObject->exists()) {
+                throw EBox::Exceptions::Internal("Unable to create $ldap{dn}");
+            }
+        } else {
+            throw EBox::Exceptions::Internal("Unable to find $ldap{dn} on LDAP.");
+        }
+    }
+    $ldbObject->_linkWithUsersObject($ldapObject);
+
+    if ($ldb{advanced}) {
+        # LDB Object is a container and API forbid to change a Container, but
+        # setting the isCriticalSystemObject is valid. Workaround the
+        # restriction.
+        my $entry = $ldbObject->_entry();
+        $entry->replace(isCriticalSystemObject => 1);
+        $entry->update($ldbObject->_ldap->connection());
+    } else {
+        my $entry = $ldbObject->_entry();
+        $entry->replace(isCriticalSystemObject => 0);
+        $entry->update($ldbObject->_ldap->connection());
+    }
+}
+
 # Method: mapDefaultContainers
 #
-#   Links the default CN containers from SAMBA in the LDAP equivalent OUs.
+#   Links the default containers between Samba LDB and LDAP, adding to the
+#   LDAP entry the msdsObjectGUID attribute, which value is the LDB object GUID
+#
+#   The mappings stablished are:
+#
+#           LDAP                                LDB
+#
+#       OU=users,$baseDN                    CN=users,$baseDN
+#       OU=computers,$baseDN                CN=computers,$baseDN
+#       OU=Builtin,$baseDN   (created)      CN=Builtin,$baseDN
+#       OU=groups,$baseDN                   OU=groups,$baseDN       (created)
+#       OU=kerberos,$baseDN                 OU=kerberos,$baseDN     (created)
 #
 sub mapDefaultContainers
 {
@@ -370,53 +447,53 @@ sub mapDefaultContainers
 
     my $usersMod = EBox::Global->modInstance('users');
     my $sambaMod = EBox::Global->modInstance('samba');
-    my @containerNames = ('Users', 'Computers', 'Builtin');
-    my $ldb = $sambaMod->ldb();
-    my $ldbRootDN = $ldb->dn();
-    my $ldapRootDN = $usersMod->ldap()->dn();
 
-    foreach my $containerName (@containerNames) {
-        my $ldbDN = "CN=$containerName,$ldbRootDN";
-        my $ldapDN = "ou=$containerName,$ldapRootDN";
+    my $ldbBaseDN = $sambaMod->ldb->dn();
+    my $ldapBaseDN = $usersMod->ldap->dn();
+    my $ldap;
+    my $ldb;
 
-        EBox::info("Mapping '$ldbDN' into '$ldapDN'");
+    # Link users container
+    $ldap = {dn => "OU=Users,$ldapBaseDN", create => 0};
+    $ldb  = {dn => "CN=Users,$ldbBaseDN", create => 0};
+    $self->linkContainer($ldb, $ldap);
 
-        my $ldbObject = $sambaMod->objectFromDN($ldbDN);
-        my $ldapObject = $usersMod->objectFromDN($ldapDN);
+    # Link computers container
+    $ldap = {dn => "OU=Computers,$ldapBaseDN", create => 0};
+    $ldb  = {dn => "CN=Computers,$ldbBaseDN", create => 0};
+    $self->linkContainer($ldb, $ldap);
 
-        unless ($ldbObject) {
-            throw EBox::Exceptions::Internal("Unable to find $ldbDN on LDB.")
-        }
+    # Link builtin container
+    $ldap = {dn => "OU=Builtin,$ldapBaseDN", create => 1};
+    $ldb  = {dn => "CN=Builtin,$ldbBaseDN", create => 0};
+    $self->linkContainer($ldb, $ldap);
 
-        if ($ldapObject) {
-            $ldbObject->_linkWithUsersObject($ldapObject);
-        } else {
-            EBox::debug("'$ldapDN' doesn't exist in LDAP, creating it...");
-            $ldbObject->addToZentyal();
-        }
-    }
+    # Link groups container
+    $ldap = {dn => "OU=Groups,$ldapBaseDN", create => 0};
+    $ldb  = {dn => "OU=Groups,$ldbBaseDN", create => 1};
+    $self->linkContainer($ldb, $ldap);
 
-    # ou=Groups is an special case, it should exists always.
-    my $ldbDN = "OU=Groups,$ldbRootDN";
-    my $ldapDN = "ou=Groups,$ldapRootDN";
-
-    EBox::info("Mapping '$ldbDN' into '$ldapDN'");
-
-    my $sambaGroupsOU = $sambaMod->objectFromDN($ldbDN);
-    my $ldapGroupsOU = $usersMod->objectFromDN($ldapDN);
-
-    unless ($sambaGroupsOU and $sambaGroupsOU->exists()) {
-        my $sambaParent = $sambaMod->defaultNamingContext();
-        $sambaGroupsOU = EBox::Samba::OU->create(name => 'Groups', parent => $sambaParent);
-    }
-
-    if ($ldapGroupsOU and $ldapGroupsOU->exists()) {
-        $sambaGroupsOU->_linkWithUsersObject($ldapGroupsOU);
-    } else {
-        $sambaGroupsOU->addToZentyal();
-    }
+    # Link kerberos container
+    $ldap = {dn => "OU=Kerberos,$ldapBaseDN", create => 0};
+    $ldb  = {dn => "OU=Kerberos,$ldbBaseDN", create => 1, advanced => 1};
+    $self->linkContainer($ldb, $ldap);
 }
 
+# Method: mapAccounts
+#
+#   Set the mapping between the objectSID and uidNumber/gidNumber for the
+#   following entries by writing these entries in the idmap database.
+#
+#   NOTE: At this point the accounts does not exists yet in LDB, but as the
+#         SIDs are well-known we can stablish the mappings
+#
+#               LDB                         System
+#       User  'Administrator'   =>      User  'root'
+#       Group 'Domain Admins'   =>      Group 'adm'
+#       Group 'Domain Users'    =>      Group '__USERS__'
+#       User  'Guest'           =>      User  'nobody'
+#       Group 'Domain Guests'   =>      Group 'nogroup'
+#
 sub mapAccounts
 {
     my ($self) = @_;
@@ -436,44 +513,16 @@ sub mapAccounts
     my $admGID = 4;
 
     EBox::info("Mapping domain administrator account");
-    my $domainAdmin = new EBox::Samba::User(sid => $domainAdminSID);
-    my $domainAdminZentyal = new EBox::Users::User(uid => $domainAdmin->get('samAccountName'));
-    if ($domainAdmin->exists()) {
-        if ($domainAdminZentyal->exists()) {
-            $domainAdmin->_linkWithUsersObject($domainAdminZentyal);
-        } else {
-            $domainAdmin->addToZentyal();
-        }
-    }
     $sambaModule->ldb->idmap->setupNameMapping($domainAdminSID, $typeUID, $rootUID);
 
     EBox::info("Mapping domain administrators group account");
-    my $domainAdmins = new EBox::Samba::Group(sid => $domainAdminsSID);
-    my $domainAdminsZentyal = new EBox::Users::Group(gid => $domainAdmins->get('samAccountName'));
-    if ($domainAdmins->exists()) {
-        if ($domainAdminsZentyal->exists()) {
-            $domainAdmins->_linkWithUsersObject($domainAdminsZentyal);
-        } else {
-            $domainAdmins->addToZentyal();
-        }
-    }
     $sambaModule->ldb->idmap->setupNameMapping($domainAdminsSID, $typeBOTH, $admGID);
 
     EBox::info("Mapping domain users group account");
-    my $usersModule = EBox::Global->modInstance('users');
     # FIXME Why is this not working during first intall???
     #my $usersGID = getpwnam($usersModule->DEFAULTGROUP());
     my $usersGID = 1901;
     my $domainUsersSID = "$domainSID-513";
-    my $domainUsers = new EBox::Samba::Group(sid => $domainUsersSID);
-    my $domainUsersZentyal = new EBox::Users::Group(gid => $usersModule->DEFAULTGROUP());
-    if ($domainUsers->exists()) {
-        if ($domainUsersZentyal->exists()) {
-            $domainUsers->_linkWithUsersObject($domainUsersZentyal);
-        } else {
-            $domainUsers->addToZentyal();
-        }
-    }
     $sambaModule->ldb->idmap->setupNameMapping($domainUsersSID, $typeGID, $usersGID);
 
     # Map domain guest account to nobody user
@@ -492,6 +541,41 @@ sub provisionDC
     my ($self, $provisionIP) = @_;
 
     my $samba = EBox::Global->modInstance('samba');
+    my $usersModule = EBox::Global->modInstance('users');
+
+    # The OU=Users will be linked to CN=Users, which can not contain OUs. Check
+    # there are not OUs created inside OU=Users.
+    my $ldap = $usersModule->ldap();
+    my $param = {
+        base => "OU=Users," . $ldap->dn(),
+        scope => 'one',
+        filter => '(objectClass=organizationalUnit)',
+        attrs => ['*']
+    };
+    my $result = $ldap->search($param);
+    if ($result->count() > 0) {
+        my $msg = __("There are nested organizational units created inside " .
+                     "the organizational unit 'Users'. This is not " .
+                     "supported and will cause the import of LDAP entries " .
+                     "to samba to fail.");
+        throw EBox::Exceptions::External($msg);
+    }
+    # Same about OU=Computers
+    $param = {
+        base => "OU=Computers," . $ldap->dn(),
+        scope => 'one',
+        filter => '(objectClass=organizationalUnit)',
+        attrs => ['*']
+    };
+    $result = $ldap->search($param);
+    if ($result->count() > 0) {
+        my $msg = __("There are nested organizational units created inside " .
+                     "the organizational unit 'Computers'. This is not " .
+                     "supported and will cause the import of LDAP entries " .
+                     "to samba to fail.");
+        throw EBox::Exceptions::External($msg);
+    }
+
     try {
         $self->setProvisioning(1);
 
@@ -499,7 +583,6 @@ sub provisionDC
 
         my $fs = EBox::Config::configkey('samba_fs');
         my $sysinfo = EBox::Global->modInstance('sysinfo');
-        my $usersModule = EBox::Global->modInstance('users');
         my $cmd = 'samba-tool domain provision ' .
             " --domain='" . $samba->workgroup() . "'" .
             " --workgroup='" . $samba->workgroup() . "'" .
@@ -576,7 +659,7 @@ sub provisionDC
     } otherwise {
         my ($error) = @_;
         $self->setProvisioned(0);
-        throw EBox::Exceptions::External($error);
+        throw EBox::Exceptions::Internal($error);
     };
 }
 
@@ -1055,6 +1138,7 @@ sub _addForestDnsZonesReplica
 {
     my ($self) = @_;
 
+    EBox::info("Adding Forest Dns replica");
     my $sambaModule = EBox::Global->modInstance('samba');
     my $ldb = $sambaModule->ldb();
     my $basedn = $ldb->dn();
@@ -1085,6 +1169,7 @@ sub _addDomainDnsZonesReplica
 {
     my ($self) = @_;
 
+    EBox::info("Adding Domain Dns replica");
     my $sambaModule = EBox::Global->modInstance('samba');
     my $ldb = $sambaModule->ldb();
     my $basedn = $ldb->dn();
@@ -1124,7 +1209,6 @@ sub _waitForRidPoolAllocation
 {
     my ($self) = @_;
 
-    use Data::Dumper; # TODO Remove
     my $allocated = 0;
     my $maxTries = 300;
     my $sleepSeconds = 0.1;
@@ -1252,13 +1336,9 @@ sub provisionADC
         EBox::debug("Setting domain DNS server '$adDnsServer' as the primary resolver");
         $dnsFile = new File::Temp(TEMPLATE => 'resolvXXXXXX',
                                   DIR      => EBox::Config::tmp());
-        EBox::Sudo::root("cp /etc/resolv.conf $dnsFile");
-        my $array = [];
-        push (@{$array}, searchDomain => $domainToJoin);
-        push (@{$array}, nameservers => [ $adDnsServer ]);
-        $sambaModule->writeConfFile(EBox::Network::RESOLV_FILE(),
-                                    'network/resolv.conf.mas',
-                                    $array);
+        EBox::Sudo::root("cp /etc/resolvconf/interface-order $dnsFile",
+                         'echo zentyal.temp > /etc/resolvconf/interface-order',
+                         "echo 'search $domainToJoin\nnameserver $adDnsServer' | resolvconf -a zentyal.temp");
 
         # Get a ticket for admin User
         my $principal = "$adUser\@$realm";
@@ -1293,12 +1373,14 @@ sub provisionADC
             }
             throw EBox::Exceptions::External("Error joining to domain: @error");
         }
-        $self->_addForestDnsZonesReplica();
-        $self->_addDomainDnsZonesReplica();
+
         $self->setupDNS();
 
         # Start managed service to let it create the LDAP socket
         $sambaModule->_startService();
+
+        $self->_addForestDnsZonesReplica();
+        $self->_addDomainDnsZonesReplica();
 
         # Wait for RID pool allocation
         EBox::info("Waiting RID pool allocation");
@@ -1329,15 +1411,39 @@ sub provisionADC
             $zentyalGroup->setIgnoredModules(['samba']);
             $zentyalGroup->deleteObject();
         }
-        foreach my $zentyalOU (@{$ous}) {
-            next if (grep { lc $_ eq lc $zentyalOU->name() } @{
-                ['kerberos', 'users', 'groups', 'computers',
-                 'postfix', 'mailalias', 'vdomains', 'fetchmail',
-                 'zarafa']
+        foreach my $zentyalOU (reverse @{$ous}) {
+            # Do not remove any OU under these bases, won't be synced to LDB
+            #   (zarafa module) OU=zarafa,$baseDN
+            #   (mail module)   OU=postfix,$baseDN
+            # Do not remove this OUs (will be mapped), but remove any child
+            #   (users module)  OU=users
+            #   (users module)  OU=groups
+            #   (users module)  OU=computers
+            #   (users module)  OU=Kerberos
+            my $dn = ldap_explode_dn($zentyalOU->dn(), reverse => 1);
+            my $rdn;
+            do {
+                $rdn = shift (@{$dn});
+            } while (scalar @{$dn} and not defined $rdn->{OU});
+            my $ouName = lc $rdn->{OU};
+
+            # Skip removal of any OU under OU=zarafa and OU=postfix
+            next if (grep { $_ eq $ouName } @{
+                [ 'postfix', 'zarafa' ]
             });
 
+            # Skip the removal of kerberos, users, groups and computers, but
+            # remove any OU inside them
+            next if (grep { $_ eq $ouName } @{
+                [ 'kerberos', 'users', 'groups', 'computers' ]
+            } and not scalar @{$dn});
+
+            # As we iterate the array in reverse order, we should not remove
+            # a parent before a child, but just in case, check for object
+            # before delete
+            $zentyalOU->clearCache();
             $zentyalOU->setIgnoredModules(['samba']);
-            $zentyalOU->deleteObject();
+            $zentyalOU->deleteObject() if $zentyalOU->exists();
         }
 
         # Map defaultContainers
@@ -1360,7 +1466,8 @@ sub provisionADC
     } finally {
         # Revert primary resolver changes
         if (defined $dnsFile and -f $dnsFile) {
-            EBox::Sudo::root("cp $dnsFile /etc/resolv.conf");
+            EBox::Sudo::root("cp $dnsFile /etc/resolvconf/interface-order",
+                             'resolvconf -d zentyal.temp');
             unlink $dnsFile;
         }
         # Remote stashed password
